Further work on new FilterProcessor impl
